requirements.txt
transformers==4.34.0
torch==2.0.1
python-dotenv==1.0.0
streamlit==1.27.2
pypdf2==3.0.1
langchain==0.0.300
google-generativeai==0.3.2
python-docx==0.8.11
nltk==3.8.1